Move repaceGreekSymbols into MathematicaParsers and remove CommonUtils.
