+++ src/ThreeHiggs/MathematicaParsers.py
@@ -1,3 +1,20 @@
+def replaceGreekSymbols(string: str) -> str:
+    #import unicodedata
+    ## Unicode magic, this is definitely not ideal
+    lowerCaseMu = u"\u03BC"
+    lowerCaseLambda = u"\u03BB"
+
+    newString = string 
+
+    newString = newString.replace(lowerCaseLambda, "lam")
+    newString = newString.replace(lowerCaseMu, "mu")
+    
+    """ TODO use unicodedata package here to do magic.
+    """
+    # NOTE: Manual replacements are definitely not a general solution. Consider problematic case: expression that contains both unicode lambda and separate symbol "lam" 
+    # So tbh I'd like to keep the symbols are they are. But parse_mathematica from sympy does not seem to manage greek symbols at all!! 
+    return newString
+
 def removeSuffices(string):
     string = string.replace("^2", "sq")
 
@@ -19,7 +36,6 @@ def parseExpression(line, remove3DSuffices = False):
         identifier, line = map(str.strip, line.split("->"))
 
     from sympy.parsing.mathematica import parse_mathematica
-    from .CommonUtils import replaceGreekSymbols
     identifier = removeSuffices(replaceGreekSymbols(identifier))
     expression = parse_mathematica(replaceGreekSymbols(line).replace("3d", ""))
     symbols = [str(symbol) for symbol in expression.free_symbols]
